проект полностью готов

+++ NotesEditor.UI/AccountWindow.xaml.cs
@@ -22,15 +22,24 @@ namespace NoteEditor.UI
     /// </summary>
     public partial class AccountWindow : Window
     {
+        private readonly INoteRepository _noteRepository;
+        private readonly ICategoryRepository _categoryRepository;
+        private readonly IPictureRepository _pictureRepository;
+        private readonly ITextRepository _textRepository;
         private readonly IUserRepository _userRepository;
         private User _currentUser {  get; set; }
-        public AccountWindow(User user, IUserRepository userRepository)
+        public AccountWindow(User user, IUserRepository userRepository, INoteRepository repository, ICategoryRepository categoryRepository,
+            IPictureRepository pictureRepository, ITextRepository textRepository)
         {
             InitializeComponent();
 
             _currentUser = user;
             DataContext = _currentUser;
             _userRepository = userRepository;
+            _noteRepository = repository;
+            _categoryRepository = categoryRepository;
+            _pictureRepository = pictureRepository;
+            _textRepository = textRepository;
 
             UpdateIsVip();
         }
@@ -67,6 +76,39 @@ namespace NoteEditor.UI
         private void DeleteAccountButton_Click(object sender, RoutedEventArgs e)
         {
             _userRepository.Delete(_currentUser);
+
+            _categoryRepository.GetAll()
+                .Where(c => c.User.Id == _currentUser.Id)
+                .ToList()
+                .ForEach(c => _categoryRepository.Delete(c));
+
+            var userNotes = _noteRepository.GetAll()
+                .Where(c => c.User.Id == _currentUser.Id)
+                .ToList();
+
+            foreach (var note in userNotes)
+            {
+                var texts = _textRepository.GetAll()
+                    .Where(t => t.Note.Id == note.Id)
+                    .ToList();
+
+                foreach (var text in texts)
+                {
+                    _textRepository.Delete(text);
+                }
+
+                var images = _pictureRepository.GetAll()
+                    .Where(i => i.Note.Id == note.Id)
+                    .ToList();
+
+                foreach (var image in images)
+                {
+                    _pictureRepository.Delete(image);
+                }
+
+                _noteRepository.Delete(note);
+            }
+
             this.DialogResult = false;
             this.Close();
         }
